Add test for rendering Runaway Info Page
This commit introduces a new test case in the `RunawayTests.razor` file to verify the rendering of the "Runaway Info Page." The test method checks that the expected HTML markup matches the output of the `Runaway` component. Additionally, a TODO comment is included to consider renaming the `Runaway` component to `RunawayYouth` in the future.